factor out makeSingleEditFieldProps and StepEditFormComponent stuff
