+++ protocol-designer/src/components/StepEditForm/forms/TemperatureForm.js
@@ -11,7 +11,6 @@ import {
   ConditionalOnField,
   TextField,
 } from '../fields'
-import { useSingleEditFieldProps } from '../fields/useSingleEditFieldProps'
 import styles from '../StepEditForm.css'
 import type { StepFormProps } from '../types'
 
@@ -23,8 +22,7 @@ export const TemperatureForm = (props: StepFormProps): React.Node => {
     uiModuleSelectors.getSingleTemperatureModuleId
   )
 
-  const propsForFields = useSingleEditFieldProps(props.focusHandlers)
-  if (propsForFields === null) return null
+  const { propsForFields } = props
 
   return (
     <div className={styles.form_wrapper}>
